Create method to pull resource list from API
Instead of using a default list of API resource types, create a simple
try/except/finally block at the beginning of the API module that
pulls the resource list and its endpoints from the API itself and builds
it into a local dictionary.

If the API call fails or returns no results, default to existing
resource list.