Update documentation to 2023 standards
Adjust documentation to be clearer in language, address changes since
the last pass through documenting the project, and adjust variables
where applicable for clarity and self-documentation.

Ties off the previous commit's work of refining class, function, method,
and variable names for clarity and brings those changes in line with
the available documentation.